# docs/swagger.json
{
  "info": {
    "title": "Zenith",
    "version": "1.0.0",
    "description": "tax calculator",
    "termsOfService": "",
    "contact": {
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/",
  "swagger": "2.0",
  "paths": {
    "/signup": {
      "post": {
        "description": "Sign up a user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "description": "Login a user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "username",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      }
    },
    "/dagration": {
      "post": {
        "description": "Adds dagration to the database",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          },
          {
            "name": "amount",
            "description": "Amount of the dagration.",
            "in": "body",
            "required": false,
            "type": "string"
          },
          {
            "name": "order",
            "description": "The order in which to apply the dagration.",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "name": "rate",
            "description": "The rate at which the dagration is calculated",
            "in": "body",
            "type": "boolean",
            "required": false
          },
          {
            "name": "force",
            "description": "Specifies whether to overwrite a dagration. A bulk insert overwrites by default",
            "in": "query",
            "type": "string",
            "required": false
          }
        ],
        "responses": {
          "201": {
            "description": "Success",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Dagrations"
              }
            }
          }
        }
      },
      "get": {
        "description": "Fetch all gradations",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Dagrations"
              }
            }
          }
        }
      }
    },
    "/calculate": {
      "get": {
        "description": "calculates tax by appling gradations in order",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "required": true,
            "type": "string",
            "default": "Bearer {token}"
          },
          {
            "name": "amount",
            "in": "query",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "success",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/Tax"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
    "NewUser": {
      "type": "object",
      "required": [
        "username",
        "password"
      ],
      "properties": {
        "username": {
          "type": "string"
        },
        "password": {
          "type": "string",
          "format": "password"
        }
      }
    },
    "User": {
      "allOf": [
        {
          "$ref": "#/definitions/NewUser"
        },
        {
          "required": [
            "id",
            "token"
          ]
        },
        {
          "properties": null,
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "token": {
            "type": "string"
          }
        }
      ]
    },
    "NewDagration": {
      "type": "object",
      "required": [
        "amount",
        "order",
        "rate"
      ],
      "properties": {
        "amount": {
          "type": "number"
        },
        "order": {
          "type": "number"
        },
        "rate": {
          "type": "number"
        }
      }
    },
    "Tax": {
      "type": "object",
      "required": [
        "tax"
      ],
      "properties": {
        "tax": {
          "type": "number"
        }
      }
    },
    "Dagrations": {
      "allOf": [
        {
          "$ref": "#/definitions/NewDagration"
        }
      ]
    }
  },
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}